fix: make update answer work properly
